# models/openAPI.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AudiMeta API",
        "version": "1.0.2",
        "description": "API to retrieve information about Book, Series and Search"
    },
    "servers": [
        {
            "url": "https://audimeta.de/",
            "description": "Public Instance"
        }
    ],
    "paths": {
...
        "/chapters/{asin}": {
            "get": {
                "tags": [
                    "chapter"
                ],
                "summary": "Get the chapters of a book",
                "parameters": [
                    {
                        "name": "region",
                        "in": "query",
                        "description": "The region to search in. Can be a comma separated list of max 3 regions. Defaults to US",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "US",
                                "UK",
                                "AU",
                                "DE",
                                "FR",
                                "CA",
                                "IT",
                                "JP",
                                "IN",
                                "BR",
                                "MX",
                                "ES"
                            ]
                        }
                    },
                    {
                        "name": "asin",
                        "in": "path",
                        "description": "The asin of the book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapters found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "asin": {
                                            "type": "string",
                                            "description": "The books asin"
                                        },
                                        "chapters": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "title": {
                                                        "type": "string",
                                                        "description": "The title of the chapter"
                                                    },
                                                    "lengthMs": {
                                                        "type": "number",
                                                        "description": "The length of the chapter in milliseconds"
                                                    },
                                                    "startOffsetMs": {
                                                        "type": "number",
                                                        "description": "The start offset of the chapter in milliseconds"
                                                    },
                                                    "startOffsetSec": {
                                                        "type": "number",
                                                        "description": "The start offset of the chapter in seconds"
                                                    }
                                                },
                                                "required": ["title", "lengthMs", "startOffsetMs", "startOffsetSec"]
                                            }
                                        },
                                        "isAccurate": {
                                            "type": "boolean",
                                            "description": "Whether the chapter information is accurate"
                                        },
                                        "runtimeLengthMs": {
                                            "type": "number",
                                            "description": "The runtime length of the book in milliseconds"
                                        },
                                        "runtimeLengthSec": {
                                            "type": "number",
                                            "description": "The runtime length of the book in seconds"
                                        }
                                    },
                                    "required": ["asin", "chapters", "isAccurate", "runtimeLengthMs", "runtimeLengthSec"]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}